+++ superstructure/generic/ComponentSpecParser/parse_setservices.F90
@@ -8,7 +8,7 @@ contains
 
    
    module function parse_setservices(config, rc) result(user_ss)
-      type(DSOSetServices) :: user_ss
+      type(DsoSetServices) :: user_ss
       type(ESMF_HConfig), target, intent(in) :: config
       integer, optional, intent(out) :: rc
 
@@ -24,10 +24,9 @@ contains
          userRoutine = 'setservices_'
       end if
 
-      user_ss = user_setservices(sharedObj, userRoutine)
+      user_ss = DsoSetServices(sharedObj, userRoutine)
       
       _RETURN(_SUCCESS)
    end function parse_setservices
 
 end submodule parse_setservices_smod
-